add .gitignore and .gitattributes

+++ compiler.h
@@ -45,12 +45,12 @@
 /* Long names  */
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
-typedef unsigned int uint32_t;
+typedef unsigned long int uint32_t;
 typedef unsigned long long uint64_t;
 
 typedef signed char sint8_t;
 typedef signed short sint16_t;
-typedef signed int sint32_t;
+typedef signed long int sint32_t;
 typedef signed long long sint64_t;
 
 /* Short names  */
